Move parsing to separate class
